[FSTORE-1265] feature_group.materialization_job.run() fails with Spark engine
