+++ python/tests/test_util.py
@@ -86,3 +86,18 @@ class TestUtil:
     def test_convert_hudi_commit_time_to_timestamp(self):
         timestamp = util.get_timestamp_from_date_string("20221118095233099")
         assert timestamp == 1668765153099
+
+    def test_get_job_url(self, mocker):
+        # Arrange
+        mock_client_get_instance = mocker.patch("hsfs.client.get_instance")
+
+        # Act
+        util.get_job_url(href="1/2/3/4/5/6/7/8")
+
+        # Assert
+        assert (
+            mock_client_get_instance.return_value.replace_public_host.call_args[0][
+                0
+            ].path
+            == "p/5/jobs/named/7/executions"
+        )
